+++ src/commands/run.rs
@@ -9,12 +9,13 @@ use std::path::PathBuf;
 use crate::chromosome;
 use crate::clustering::{self, ClusteringConfig};
 use crate::constants::{
-    self, EnergyWeights, CONVERGENCE_THRESHOLD, CONVERGENCE_WINDOW, ENABLE_EARLY_STOPPING,
-    MAX_GENERATIONS, POPULATION_SIZE,
+    self, EnergyWeights, CONVERGENCE_THRESHOLD, CONVERGENCE_WINDOW, MAX_GENERATIONS,
+    POPULATION_SIZE,
 };
 use crate::evaluator;
 use crate::population;
 use crate::restraints;
+use crate::runner::run_ga;
 use crate::scoring;
 use crate::structure::{self, read_pdb, Molecule};
 use crate::utils;
@@ -34,178 +35,7 @@ pub struct RunConfig {
 /// Re-exported for use by tests and other modules that imported from here.
 pub use crate::structure::combine_molecules;
 
-// ============================================================================
-// Hall of Fame - tracks diverse solutions throughout evolution
-// ============================================================================
-
-/// Default parameters for Hall of Fame
-const HALL_OF_FAME_MAX_SIZE: usize = 500;
-const HALL_OF_FAME_TOP_K: usize = 10;
-const UNIQUENESS_ROTATION_THRESHOLD: f64 = 0.2; // ~11 degrees
-const UNIQUENESS_TRANSLATION_THRESHOLD: f64 = 2.0; // 2 Å
-
-/// Number of output models after clustering
 const NUM_OUTPUT_MODELS: usize = 5;
-
-/// Entry in the Hall of Fame storing a solution's genes and metadata
-#[derive(Debug, Clone)]
-pub struct HallOfFameEntry {
-    /// Chromosome genes: [alpha, beta, gamma, dx, dy, dz]
-    pub genes: [f64; 6],
-    /// Fitness score at time of capture
-    pub fitness: f64,
-    /// Generation when captured
-    pub generation: u32,
-}
-
-/// Tracks diverse high-quality solutions throughout evolution
-#[derive(Debug)]
-pub struct HallOfFame {
-    entries: Vec<HallOfFameEntry>,
-    max_size: usize,
-}
-
-impl HallOfFame {
-    /// Create a new Hall of Fame with default max size
-    pub fn new() -> Self {
-        Self {
-            entries: Vec::new(),
-            max_size: HALL_OF_FAME_MAX_SIZE,
-        }
-    }
-
-    /// Create a new Hall of Fame with custom max size
-    pub fn with_capacity(max_size: usize) -> Self {
-        Self {
-            entries: Vec::new(),
-            max_size,
-        }
-    }
-
-    /// Try to add a new entry if it's unique enough
-    /// Returns true if the entry was added
-    pub fn try_add(&mut self, genes: &[f64], fitness: f64, generation: u32) -> bool {
-        if genes.len() != 6 {
-            return false;
-        }
-
-        let new_genes: [f64; 6] = [genes[0], genes[1], genes[2], genes[3], genes[4], genes[5]];
-
-        // Check uniqueness against existing entries
-        if !self.is_unique(&new_genes) {
-            return false;
-        }
-
-        // Add the entry
-        self.entries.push(HallOfFameEntry {
-            genes: new_genes,
-            fitness,
-            generation,
-        });
-
-        // If over capacity, remove worst entries
-        if self.entries.len() > self.max_size {
-            self.prune();
-        }
-
-        true
-    }
-
-    /// Check if genes are unique compared to existing entries
-    fn is_unique(&self, new_genes: &[f64; 6]) -> bool {
-        for entry in &self.entries {
-            if Self::genes_are_similar(new_genes, &entry.genes) {
-                return false;
-            }
-        }
-        true
-    }
-
-    /// Check if two gene sets are similar (potential duplicates)
-    fn genes_are_similar(a: &[f64; 6], b: &[f64; 6]) -> bool {
-        // Check rotations (with wrap-around handling)
-        for i in 0..3 {
-            let diff = Self::angular_difference(a[i], b[i]);
-            if diff > UNIQUENESS_ROTATION_THRESHOLD {
-                return false;
-            }
-        }
-
-        // Check translations
-        for i in 3..6 {
-            let diff = (a[i] - b[i]).abs();
-            if diff > UNIQUENESS_TRANSLATION_THRESHOLD {
-                return false;
-            }
-        }
-
-        true
-    }
-
-    /// Calculate minimum angular difference (handles wrap-around at 2π)
-    fn angular_difference(a: f64, b: f64) -> f64 {
-        use std::f64::consts::PI;
-        let diff = (a - b).abs();
-        if diff > PI {
-            2.0 * PI - diff
-        } else {
-            diff
-        }
-    }
-
-    /// Remove worst entries when over capacity
-    fn prune(&mut self) {
-        // Sort by fitness (lower is better) and keep best
-        self.entries
-            .sort_by(|a, b| a.fitness.partial_cmp(&b.fitness).unwrap());
-        self.entries.truncate(self.max_size);
-    }
-
-    /// Get the number of entries
-    pub fn len(&self) -> usize {
-        self.entries.len()
-    }
-
-    /// Check if empty
-    pub fn is_empty(&self) -> bool {
-        self.entries.is_empty()
-    }
-
-    /// Get all entries (for clustering)
-    pub fn entries(&self) -> &[HallOfFameEntry] {
-        &self.entries
-    }
-
-    /// Add top-K chromosomes from a population
-    pub fn add_from_population(&mut self, chromosomes: &[chromosome::Chromosome], generation: u32) {
-        // Get indices sorted by fitness (best first)
-        let mut indices: Vec<usize> = (0..chromosomes.len()).collect();
-        indices.sort_by(|&a, &b| {
-            chromosomes[a]
-                .fitness
-                .partial_cmp(&chromosomes[b].fitness)
-                .unwrap()
-        });
-
-        // Try to add top-K
-        let mut added = 0;
-        for &idx in &indices {
-            if added >= HALL_OF_FAME_TOP_K {
-                break;
-            }
-            let chr = &chromosomes[idx];
-            if self.try_add(&chr.genes, chr.fitness, generation) {
-                added += 1;
-            }
-        }
-    }
-}
-
-impl Default for HallOfFame {
-    fn default() -> Self {
-        Self::new()
-    }
-}
 
 // ============================================================================
 
@@ -351,25 +181,15 @@ pub fn run(config: RunConfig) {
     }
 
     // Evolve the population
-    let mut generation_count = 0u64;
     let mut best_score_history: Vec<f64> = Vec::new();
-    let mut generations_without_improvement = 0;
-    let mut last_best_score = f64::MAX;
-
-    // Initialize Hall of Fame for tracking diverse solutions
-    let mut hall_of_fame = HallOfFame::new();
 
     println!(
         "{} Starting evolution for {} generations",
         "🧬".bold(),
         MAX_GENERATIONS
     );
-    while generation_count < MAX_GENERATIONS {
-        pop.eval_fitness();
 
-        // Track diverse solutions in Hall of Fame
-        hall_of_fame.add_from_population(&pop.chromosomes, generation_count as u32);
-
+    let ga_result = run_ga(pop, &mut rng, MAX_GENERATIONS, |gen, pop| {
         // Calculate metrics for all chromosomes (only if reference is available)
         let metric_vec = eval.as_ref().map(|e| pop.eval_metrics(e));
 
@@ -385,30 +205,14 @@ pub fn run(config: RunConfig) {
         let best_fitness = pop.chromosomes[best_fitness_idx].fitness;
         let best_chr = &pop.chromosomes[best_fitness_idx];
 
-        // Track convergence
         best_score_history.push(best_fitness);
-        if generation_count > 0 {
-            let improvement = (last_best_score - best_fitness) / last_best_score.abs();
-            if improvement < CONVERGENCE_THRESHOLD {
-                generations_without_improvement += 1;
-            } else {
-                generations_without_improvement = 0;
-            }
-        }
-        last_best_score = best_fitness;
 
-        // Check for early stopping
-        if ENABLE_EARLY_STOPPING && generations_without_improvement >= CONVERGENCE_WINDOW {
-            progress.finish_with_message(format!(
-                "{} Converged at generation {} (no improvement larger than {}% for {} gens)",
-                "✓".green(),
-                generation_count,
-                CONVERGENCE_THRESHOLD * 100.0,
-                CONVERGENCE_WINDOW
-            ));
-            println!();
-            break;
-        }
+        let improvement_since_last = if gen > 0 {
+            let prev = best_score_history[gen as usize - 1];
+            ((prev - best_fitness) / prev.abs()) * 100.0
+        } else {
+            0.0
+        };
 
         // Calculate population averages
         let mean_fitness = pop.get_mean_fitness();
@@ -419,19 +223,10 @@ pub fn run(config: RunConfig) {
                 .map(|c| 1.0 - c.restraint_penalty / num_restraints as f64)
                 .sum::<f64>()
                 / pop.chromosomes.len() as f64);
-
         let best_rest = 100.0 * (1.0 - best_chr.restraint_penalty / num_restraints as f64);
 
-        // Calculate improvement metrics
-        let improvement_since_last = if generation_count > 0 {
-            let prev = best_score_history[generation_count as usize - 1];
-            ((prev - best_fitness) / prev.abs()) * 100.0
-        } else {
-            0.0
-        };
-
         // Update progress bar and print output
-        progress.set_position(generation_count);
+        progress.set_position(gen);
 
         if let Some(ref metrics) = metric_vec {
             let mean_dockq: f64 =
@@ -470,7 +265,7 @@ pub fn run(config: RunConfig) {
                 best_fitness
             };
             progress.println(format!("  [{}] {} score={:>8.3} dockq={} rest={}% │ {} score={:>8.3} dockq={} rest={}% rmsd={:.2}Å fnat={:.3} irmsd={:.2}Å │ Δ={}%",
-                format!("{:>3}", generation_count).bright_black(),
+                format!("{:>3}", gen).bright_black(),
                 "📊".bright_blue(),
                 mean_score_display,
                 format!("{:.3}", mean_dockq).cyan(),
@@ -497,7 +292,7 @@ pub fn run(config: RunConfig) {
             progress.set_message(format!("Score: {:.0}", best_fitness));
             progress.println(format!(
                 "  [{}] {} score={:>8.1} rest={}% │ {} score={:>8.1} rest={}% │ Δ={}%",
-                format!("{:>3}", generation_count).bright_black(),
+                format!("{:>3}", gen).bright_black(),
                 "📊".bright_blue(),
                 mean_fitness,
                 format!("{:>3.0}", mean_rest).bright_black(),
@@ -511,15 +306,24 @@ pub fn run(config: RunConfig) {
                 }
             ));
         }
+    });
 
-        generation_count += 1;
-        pop = pop.evolve(&mut rng);
-    }
-
-    // Finish progress bar if not already finished
-    if !progress.is_finished() {
+    // Finish progress bar
+    if ga_result.converged_early {
+        progress.finish_with_message(format!(
+            "{} Converged at generation {} (no improvement larger than {}% for {} gens)",
+            "✓".green(),
+            ga_result.generations_run,
+            CONVERGENCE_THRESHOLD * 100.0,
+            CONVERGENCE_WINDOW
+        ));
+        println!();
+    } else {
         progress.finish();
     }
+
+    let hall_of_fame = ga_result.hall_of_fame;
+    let pop = ga_result.final_population;
 
     // Determine output directory
     let out_dir = match &output_dir {
@@ -535,9 +339,6 @@ pub fn run(config: RunConfig) {
         // =====================================================================
         // No clustering: Output best_by_score and best_by_dockq (old behavior)
         // =====================================================================
-
-        // Final evaluation to get best models from final population
-        pop.eval_fitness();
 
         let best_fitness_idx = pop
             .chromosomes
@@ -878,7 +679,6 @@ pub fn run(config: RunConfig) {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::f64::consts::PI;
 
     #[test]
     fn test_combine_molecules() {
@@ -895,126 +695,5 @@ mod tests {
             receptor.0.len() + ligand.0.len(),
             "Combined molecule should have all atoms from both"
         );
-    }
-
-    #[test]
-    fn test_hall_of_fame_new() {
-        let hof = HallOfFame::new();
-        assert!(hof.is_empty());
-        assert_eq!(hof.len(), 0);
-    }
-
-    #[test]
-    fn test_hall_of_fame_add_entry() {
-        let mut hof = HallOfFame::new();
-        let genes = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0];
-
-        let added = hof.try_add(&genes, -100.0, 1);
-        assert!(added);
-        assert_eq!(hof.len(), 1);
-    }
-
-    #[test]
-    fn test_hall_of_fame_rejects_duplicates() {
-        let mut hof = HallOfFame::new();
-        let genes = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0];
-
-        hof.try_add(&genes, -100.0, 1);
-        let added = hof.try_add(&genes, -100.0, 2);
-
-        assert!(!added, "Should reject identical genes");
-        assert_eq!(hof.len(), 1);
-    }
-
-    #[test]
-    fn test_hall_of_fame_accepts_different_genes() {
-        let mut hof = HallOfFame::new();
-
-        // Different rotations (large difference)
-        let genes1 = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0];
-        let genes2 = [PI, PI, PI, 0.0, 0.0, 0.0];
-
-        hof.try_add(&genes1, -100.0, 1);
-        let added = hof.try_add(&genes2, -90.0, 2);
-
-        assert!(added, "Should accept different genes");
-        assert_eq!(hof.len(), 2);
-    }
-
-    #[test]
-    fn test_hall_of_fame_accepts_different_translations() {
-        let mut hof = HallOfFame::new();
-
-        // Same rotation, different translation
-        let genes1 = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0];
-        let genes2 = [0.0, 0.0, 0.0, 10.0, 10.0, 10.0];
-
-        hof.try_add(&genes1, -100.0, 1);
-        let added = hof.try_add(&genes2, -90.0, 2);
-
-        assert!(added, "Should accept different translations");
-        assert_eq!(hof.len(), 2);
-    }
-
-    #[test]
-    fn test_hall_of_fame_rejects_similar_genes() {
-        let mut hof = HallOfFame::new();
-
-        // Very similar genes (within threshold)
-        let genes1 = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0];
-        let genes2 = [0.1, 0.1, 0.1, 0.5, 0.5, 0.5]; // Within thresholds
-
-        hof.try_add(&genes1, -100.0, 1);
-        let added = hof.try_add(&genes2, -90.0, 2);
-
-        assert!(!added, "Should reject similar genes within threshold");
-        assert_eq!(hof.len(), 1);
-    }
-
-    #[test]
-    fn test_hall_of_fame_prune_keeps_best() {
-        let mut hof = HallOfFame::with_capacity(3);
-
-        // Add 5 entries with different genes
-        let base_translation = [0.0, 0.0, 0.0];
-        for i in 0..5 {
-            let genes = [
-                (i as f64) * 1.0, // Vary rotation significantly
-                0.0,
-                0.0,
-                base_translation[0] + (i as f64) * 10.0, // Vary translation
-                base_translation[1],
-                base_translation[2],
-            ];
-            hof.try_add(&genes, -(100.0 - i as f64 * 10.0), i as u32);
-        }
-
-        // Should have pruned to max_size, keeping best fitness
-        assert_eq!(hof.len(), 3);
-
-        // Best entries should be kept (lowest fitness)
-        let entries = hof.entries();
-        assert!(entries[0].fitness <= entries[1].fitness);
-        assert!(entries[1].fitness <= entries[2].fitness);
-    }
-
-    #[test]
-    fn test_angular_difference_same() {
-        let diff = HallOfFame::angular_difference(0.0, 0.0);
-        assert!((diff - 0.0).abs() < 0.001);
-    }
-
-    #[test]
-    fn test_angular_difference_wrap_around() {
-        // 0 and 2π should be very close
-        let diff = HallOfFame::angular_difference(0.0, 2.0 * PI - 0.1);
-        assert!(diff < 0.2, "Should handle wrap-around");
-    }
-
-    #[test]
-    fn test_angular_difference_opposite() {
-        // 0 and π should be π apart
-        let diff = HallOfFame::angular_difference(0.0, PI);
-        assert!((diff - PI).abs() < 0.001);
     }
 }
