+++ src/lib.rs
@@ -4,8 +4,10 @@ pub mod constants;
 pub mod evaluator;
 pub mod fitness;
 pub mod ga;
+pub mod hall_of_fame;
 pub mod population;
 pub mod restraints;
+pub mod runner;
 pub mod structure;
 pub mod toppar;
 pub mod utils;
